feat: networkModule 추가 (#7)
